Move LogoutUtils to utility module
